fix(bot): move artist discography helper out of commands scan path
getCommandsFromDirectory flat-scans every .ts/.js file directly under
commands/ and expects a default/command export shaped {data, execute}.
The discography helper exports only a named handler function used by
artist.ts, so every boot logged:
  ERROR Command in artistDiscography.js is not a valid Command instance

Moved it into commands/artist/ (same basename as artist.ts, so the
loader's flatBaseNames check skips the directory too), matching how
autoplay/ and play/ already keep helpers out of the flat command scan.
